fix: remove utils from sections package
