+++ packages/ot-constants/src/types/response.ts
@@ -1,0 +1,5 @@
+export type responseType = {
+  data: Record<string, unknown> | null;
+  error: Record<string, unknown> | null;
+  loading: boolean;
+};
