+++ packages/ot-utils/src/index.ts
@@ -4,3 +4,4 @@ export * from "./interactors";
 export * from "./global";
 export * from "./urls";
 export * from "./comparators";
+export * from "./study";
