+++ packages/sections/src/credibleSet/Variants/Body.tsx
@@ -11,8 +11,7 @@ import { naLabel, initialResponse, table5HChunkSize } from "../../constants";
 import { definition } from ".";
 import Description from "./Description";
 import VARIANTS_QUERY from "./VariantsQuery.gql";
-import { mantissaExponentComparator, variantComparator } from "@ot/utils";
-import { identifiersOrgLink } from "../../utils/global";
+import { mantissaExponentComparator, variantComparator, identifiersOrgLink } from "@ot/utils";
 import { useEffect, useState } from "react";
 
 type getColumnsType = {
