+++ packages/sections/src/drug/Pharmacogenomics/Body.jsx
@@ -3,12 +3,11 @@ import { useQuery } from "@apollo/client";
 import { makeStyles } from "@mui/styles";
 import { Link, SectionItem, Tooltip, LabelChip, PublicationsDrawer, OtTable } from "ui";
 
-import { epmcUrl } from "../../utils/urls";
+import { epmcUrl, identifiersOrgLink, sentenceCase } from "@ot/utils";
 import { definition } from ".";
 import Description from "./Description";
 import PHARMACOGENOMICS_QUERY from "./Pharmacogenomics.gql";
 import { naLabel, PHARM_GKB_COLOR, variantConsequenceSource } from "../../constants";
-import { identifiersOrgLink, sentenceCase } from "../../utils/global";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faCircleCheck } from "@fortawesome/free-solid-svg-icons";
 import { faCircleXmark } from "@fortawesome/free-regular-svg-icons";
