+++ packages/sections/src/evidence/CancerBiomarkers/BiomarkersDrawer.jsx
@@ -1,11 +1,10 @@
 import { useState } from "react";
-import { Drawer, Link as MuiLink, IconButton, Paper, Typography, ButtonBase } from "@mui/material";
+import { Drawer, IconButton, Paper, Typography, ButtonBase } from "@mui/material";
 import { makeStyles } from "@mui/styles";
 import { faXmark } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { Tooltip, Link } from "ui";
-
-import { sentenceCase } from "../../utils/global";
+import { sentenceCase } from "@ot/utils";
 
 const useStyles = makeStyles(theme => ({
   drawerLink: {
