+++ packages/sections/src/evidence/CancerGeneCensus/Body.jsx
@@ -5,11 +5,10 @@ import { v1 } from "uuid";
 
 import { ChipList, Link, SectionItem, PublicationsDrawer, OtTable } from "ui";
 
-import { naLabel, defaultRowsPerPageOptions, sectionsBaseSizeQuery } from "../../constants";
+import { naLabel, sectionsBaseSizeQuery } from "../../constants";
 import { dataTypesMap } from "@ot/constants";
 import Description from "./Description";
-import { epmcUrl } from "../../utils/urls";
-import { identifiersOrgLink, sentenceCase } from "../../utils/global";
+import { epmcUrl, identifiersOrgLink, sentenceCase } from "@ot/utils";
 
 import { definition } from ".";
