+++ packages/sections/src/evidence/GeneBurden/Body.jsx
@@ -13,7 +13,7 @@ import {
 
 import { definition } from ".";
 import Description from "./Description";
-import { epmcUrl } from "../../utils/urls";
+import { epmcUrl } from "@ot/utils";
 import { dataTypesMap } from "@ot/constants";
 import { naLabel, sectionsBaseSizeQuery } from "../../constants";
