+++ packages/sections/src/evidence/Impc/Body.jsx
@@ -15,7 +15,7 @@ import { definition } from ".";
 import Description from "./Description";
 import INTOGEN_QUERY from "./sectionQuery.gql";
 import { dataTypesMap } from "@ot/constants";
-import { sentenceCase } from "../../utils/global";
+import { sentenceCase } from "@ot/utils";
 import { naLabel } from "../../constants";
 import { useState } from "react";
