+++ packages/sections/src/evidence/IntOgen/Body.jsx
@@ -7,10 +7,9 @@ import { ChipList, Link, SectionItem, Tooltip, ScientificNotation, OtTable } fro
 import { definition } from ".";
 import methods from "./methods";
 import Description from "./Description";
-import { epmcUrl } from "../../utils/urls";
+import { epmcUrl, sentenceCase } from "@ot/utils";
 import { dataTypesMap } from "@ot/constants";
 import INTOGEN_QUERY from "./sectionQuery.gql";
-import { sentenceCase } from "../../utils/global";
 import { naLabel, sectionsBaseSizeQuery } from "../../constants";
 
 const intOgenUrl = (id, approvedSymbol) =>
