+++ packages/sections/src/evidence/Reactome/Body.jsx
@@ -12,10 +12,9 @@ import {
 
 import { definition } from ".";
 import Description from "./Description";
-import { epmcUrl } from "../../utils/urls";
+import { epmcUrl, sentenceCase } from "@ot/utils";
 import { dataTypesMap } from "@ot/constants";
 import REACTOME_QUERY from "./sectionQuery.gql";
-import { sentenceCase } from "../../utils/global";
 import { defaultRowsPerPageOptions, naLabel, sectionsBaseSizeQuery } from "../../constants";
 
 const getColumns = label => [
