+++ packages/sections/src/evidence/SysBio/Body.jsx
@@ -3,7 +3,7 @@ import { Link, SectionItem, Tooltip, PublicationsDrawer, OtTable } from "ui";
 
 import { definition } from ".";
 import Description from "./Description";
-import { epmcUrl } from "../../utils/urls";
+import { epmcUrl } from "@ot/utils";
 import SYSBIO_QUERY from "./sectionQuery.gql";
 import { dataTypesMap } from "@ot/constants";
 import { naLabel, sectionsBaseSizeQuery } from "../../constants";
