+++ packages/sections/src/evidence/UniProtLiterature/Body.jsx
@@ -4,11 +4,10 @@ import { Link, SectionItem, Tooltip, PublicationsDrawer, OtTable } from "ui";
 
 import { definition } from ".";
 import Description from "./Description";
-import { epmcUrl } from "../../utils/urls";
+import { epmcUrl, identifiersOrgLink, sentenceCase } from "@ot/utils";
 import { dataTypesMap } from "@ot/constants";
 import { sectionsBaseSizeQuery } from "../../constants";
 import UNIPROT_LITERATURE_QUERY from "./UniprotLiteratureQuery.gql";
-import { identifiersOrgLink, sentenceCase } from "../../utils/global";
 
 const getcolumns = label => [
   {
