+++ packages/sections/src/study/GWASCredibleSets/Body.tsx
@@ -11,14 +11,14 @@ import {
   Navigate,
 } from "ui";
 import { Box } from "@mui/material";
+import { mantissaExponentComparator, nullishComparator, variantComparator } from "@ot/utils";
+import { responseType } from "@ot/constants";
 import { naLabel, credsetConfidenceMap, initialResponse, table5HChunkSize } from "../../constants";
 import { definition } from ".";
 import Description from "./Description";
 import GWAS_CREDIBLE_SETS_QUERY from "./GWASCredibleSetsQuery.gql";
-import { mantissaExponentComparator, nullishComparator, variantComparator } from "@ot/utils";
 import ManhattanPlot from "./ManhattanPlot";
 import { ReactElement, useEffect, useState } from "react";
-import { responseType } from "ui/src/types/response";
 
 const columns = [
   {
