+++ packages/sections/src/variant/Pharmacogenomics/Body.tsx
@@ -6,7 +6,7 @@ import { Fragment } from "react";
 import classNames from "classnames";
 import { makeStyles } from "@mui/styles";
 import { Link, Tooltip, PublicationsDrawer, OtTable, SectionItem } from "ui";
-import { epmcUrl } from "../../utils/urls";
+import { epmcUrl } from "@ot/utils";
 import { naLabel, PHARM_GKB_COLOR } from "../../constants";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faCircleCheck } from "@fortawesome/free-solid-svg-icons";
