+++ packages/sections/src/variant/QTLCredibleSets/Body.tsx
@@ -15,8 +15,8 @@ import { definition } from ".";
 import Description from "./Description";
 import QTL_CREDIBLE_SETS_QUERY from "./QTLCredibleSetsQuery.gql";
 import { mantissaExponentComparator, variantComparator } from "@ot/utils";
+import { responseType } from "@ot/constants";
 import { ReactNode, useEffect, useState } from "react";
-import { responseType } from "ui/src/types/response";
 
 type getColumnsType = {
   id: string;
