+++ packages/sections/src/variant/UniProtVariants/Body.tsx
@@ -3,7 +3,7 @@ import { Typography } from "@mui/material";
 import { Link, SectionItem, Tooltip, PublicationsDrawer, OtTable } from "ui";
 import { definition } from ".";
 import Description from "./Description";
-import { epmcUrl } from "../../utils/urls";
+import { epmcUrl } from "@ot/utils";
 import { naLabel } from "../../constants";
 import UNIPROT_VARIANTS_QUERY from "./UniProtVariantsQuery.gql";
