+++ packages/sections/src/variant/VariantEffectPredictor/Body.tsx
@@ -5,7 +5,7 @@ import { Fragment } from "react";
 import { definition } from "../VariantEffectPredictor";
 import Description from "../VariantEffectPredictor/Description";
 import { naLabel } from "../../constants";
-import { identifiersOrgLink } from "../../utils/global";
+import { identifiersOrgLink } from "@ot/utils";
 import VARIANT_EFFECT_PREDICTOR_QUERY from "./VariantEffectPredictorQuery.gql";
 
 function formatVariantConsequenceLabel(label) {
